Provisional removal of the progress bar
The progress bar works fine, but creates difficulties when the jQuery
animate method is used for the functionality of the "play" buttons on
the hoteltable element.  The progress bar will thus be removed but only
provisionally.  In later commits, the progress bar may be re-included if
a workaround is found for the conflict with the jQuery animate
directives.